+++ boltprojection/handler_test.go
@@ -1,10 +1,10 @@
-package sql_test
+package boltprojection_test
 
 import (
 	"context"
 	"testing"
 
-	. "github.com/dogmatiq/projectionkit/sql"
+	. "github.com/dogmatiq/projectionkit/boltprojection"
 )
 
 func TestNoCompactBehavior_Compact_ReturnsNil(t *testing.T) {